+++ drizzle-supporter.config.ts
@@ -7,9 +7,8 @@ const databaseUrl = process.env.DATABASE_SUPPORTER_URL || "";
 export default {
   schema: "./src/db/supporter/schema",
   out: "./drizzle/supporter",
-  dialect: "turso",
+  dialect: "postgresql",
   dbCredentials: {
     url: databaseUrl,
-    authToken: process.env.DATABASE_WRITE_TOKEN,
   },
 } satisfies Config;